requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.2
Pillow==8.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.2.0
pyOpenSSL==21.0.0
python-dotenv==0.19.1
pytz==2021.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
zope.interface==5.4.0
